Increase the value of skewed_partition_min_data_processed_rebalance_threshold
This session property specifies the minimum
data processed to trigger skewed partition
rebalancing in local and remote exchange.

Currently, the value of this property is too
small and since now we do not work with
physicalWrittenBytes in scaling, we can increase
the value of this property. This will also make
scaling a bit less aggressive.